Add app.rb file and refactir main.rb code
